Removed .DS_Store from tracking and added to .gitignore
