+++ fapi/main.py
@@ -1,22 +1,16 @@
-
 # wbl-backend/fapi/main.py
-from fapi.models import EmailRequest, UserCreate, Token, UserRegistration, ContactForm, ResetPasswordRequest, ResetPassword ,GoogleUserCreate, VendorCreate , RecentPlacement , RecentInterview,CandidateMarketing,Candidate, CandidateCreate, CandidateUpdate,LeadCreate,Lead,CandidatePlacement, CandidatePlacementCreate,CandidateMarketingCreate
-from  fapi.db import (
+from fapi.db.models import EmailRequest, UserCreate, Token, UserRegistration, ContactForm, ResetPasswordRequest, ResetPassword ,GoogleUserCreate, VendorCreate , RecentPlacement , RecentInterview,CandidateMarketing,Candidate, CandidateCreate, CandidateUpdate,LeadCreate,Lead,CandidatePlacement, CandidatePlacementCreate,CandidateMarketingCreate, TalentSearch
+from  fapi.db.database import (
       fetch_sessions_by_type, fetch_types,fetch_candidates, insert_login_history, insert_user, get_user_by_username, update_login_info, verify_md5_hash,
     fetch_keyword_recordings, fetch_keyword_presentation,fetch_interviews_by_name,insert_interview,delete_interview,update_interview,
- fetch_course_batches, fetch_subject_batch_recording, user_contact, course_content, fetch_candidate_id_by_email,get_candidates_by_status,fetch_interview_by_id,
-    unsubscribe_user, update_user_password ,get_user_by_username, update_user_password ,insert_user,get_google_user_by_email,insert_google_user_db,fetch_candidate_id_by_email,insert_vendor ,fetch_recent_placements , fetch_recent_interviews, get_candidate_by_name, get_candidate_by_id, create_candidate, delete_candidate as db_delete_candidate,update_candidate as db_update_candidate,get_all_placements,
-    get_placement_by_id,SessionLocal,engine,update_placement,delete_placement,get_connection
-
+ fetch_course_batches, fetch_subject_batch_recording, user_contact, course_content, fetch_candidate_id_by_email,fetch_interview_by_id,
+    unsubscribe_user, update_user_password ,get_user_by_username, update_user_password ,insert_user,get_google_user_by_email,insert_google_user_db,fetch_candidate_id_by_email,insert_vendor ,fetch_recent_placements , fetch_recent_interviews,insert_lead_new
 )
-from fapi.schemas import LeadORM,Base
-from .utils.lead_utils import fetch_all_leads_paginated, get_lead_by_id,create_lead,update_lead,delete_lead
-# , create_lead, update_lead, delete_lead
 from typing import Dict, Any
-from  fapi.auth_utils import md5_hash, verify_md5_hash, create_reset_token, verify_reset_token
+from  fapi.utils.auth_utils import md5_hash, verify_md5_hash, create_reset_token, verify_reset_token
 from  fapi.auth import create_access_token, verify_token, JWTAuthorizationMiddleware, generate_password_reset_token, verify_password_reset_token, get_password_hash ,create_google_access_token,determine_user_role
-from  fapi.contactMailTemplet import ContactMail_HTML_templete
-from  fapi.mail_service import send_reset_password_email ,send_request_demo_emails
+from  fapi.mail.templets.contactMailTemplet import ContactMail_HTML_templete
+from  fapi.utils.email_utils import send_reset_password_email ,send_request_demo_emails,send_contact_emails,send_email_to_user
 from fastapi import FastAPI, Depends, HTTPException, Request, status, Query, Body ,APIRouter, status as http_status,Path
 from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm,HTTPBearer, HTTPAuthorizationCredentials
 from fastapi.middleware.cors import CORSMiddleware
@@ -32,41 +26,13 @@ from email.mime.multipart import MIMEMultipart
 from datetime import date,datetime, timedelta
 import jwt
 from sqlalchemy.orm import Session
-
-
-from fapi.utils.candidate_utils import (
-    get_all_placements,
-    get_placement_by_id,
-    create_placement,
-    update_placement,
-    delete_placement,
-)
-
-
-from fapi.utils.candidate_utils import (
-    get_all_marketing_records,
-    get_marketing_by_id,
-    create_marketing,
-    update_marketing,
-    delete_marketing,
-)
-
-
+from fapi.db.database import Base, engine
+from fapi.api.routes import candidate, leads
 from fastapi import Query, Path
-
-from fapi.db import (
-    # get_all_candidates,
-    get_candidate_by_id,
-    create_candidate,
-    update_candidate,
-    delete_candidate,
-    get_all_candidates_paginated
-)
-
 from slowapi import Limiter
 from slowapi.util import get_remote_address
-from fapi.models import VendorResponse
-from fapi.db import db_config
+from fapi.db.models import VendorResponse
+from fapi.db.database import db_config
 from typing import Dict, Any
 from fastapi import FastAPI, Query, Path
 
@@ -74,8 +40,11 @@ load_dotenv()
 
 Base.metadata.create_all(bind=engine)
 app = FastAPI()
+app.include_router(candidate.router, prefix="/api", tags=["Candidate Marketing & Placements"])
+app.include_router(leads.router, prefix="/api", tags=["Leads"])
+
 def get_db():
-    db = SessionLocal()
+    db.database = SessionLocal()
     try:
         yield db
     finally:
@@ -149,191 +118,8 @@ def determine_user_role(userinfo):
         return "admin"
 
     return "candidate"
-# ------------------------------------------------- Candidate ------------------------------------------
 
-#GET all candidates
-@app.get("/api/candidates", response_model=List[Candidate])
-async def get_all_candidates_endpoint(page: int = 1, limit: int = 100):
-    try:
-        rows = get_all_candidates_paginated(page, limit)
-        return [Candidate(**row) for row in rows]
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-    
-    
-    # GET Candidate status
-@app.get("/api/candidates/{status}", response_model=List[Candidate])
-async def get_candidates_by_dynamic_status(
-    status: str,
-    page: int = 1,
-    limit: int = 100
-):
-    # Define valid statuses (you can expand this set as needed)
-    valid_statuses = {"active", "marketing"}  
-
-    if status.lower() not in valid_statuses:
-        raise HTTPException(status_code=400, detail=f"Invalid status: {status}")
-
-    try:
-        rows = get_candidates_by_status(status, page, limit)
-        return [Candidate(**row) for row in rows]
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-    
-    
-
-
-#GET candidate by Name
-@app.get("/api/candidates/by-name/{name}", response_model=List[Candidate])
-async def get_candidates_by_name_endpoint(name: str):
-    candidates = get_candidate_by_name(name)
-    if not candidates:
-        raise HTTPException(status_code=404, detail="Candidate not found")
-    return candidates
-
-
-# GET candidate by ID
-@app.get("/api/candidates/{candidateid}", response_model=Candidate)
-async def get_candidate(candidateid: int):
-    candidate = get_candidate_by_id(candidateid)
-    if not candidate:
-        raise HTTPException(status_code=404, detail="Candidate not found")
-    return Candidate(**candidate)
-
-
-    
-
-
-
-
-# POST - Create candidate
-@app.post("/api/candidates", response_model=Candidate)
-async def create_candidate_endpoint(candidate: CandidateCreate):
-    try:
-        fields = candidate.dict(exclude_unset=True)
-        new_id = create_candidate(fields)
-        return Candidate(**fields, candidateid=new_id)
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Insertion failed: {str(e)}")
-    
-
-# PUT - Update candidate
-@app.put("/api/candidates/{candidateid}", response_model=Candidate)
-async def update_candidate(candidateid: int, update_data: CandidateUpdate):
-    fields = update_data.dict(exclude_unset=True)
-    if not fields:
-        raise HTTPException(status_code=400, detail="No data to update")
-    try:
-        db_update_candidate(candidateid, fields)
-        return Candidate(**fields, candidateid=candidateid)
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Update failed: {str(e)}")
-
-
-# DELETE candidate
-@app.delete("/api/candidates/{candidateid}")
-async def delete_candidate(candidateid: int):
-    try:
-        # You can enhance db.py to return rowcount if needed
-        db_delete_candidate(candidateid)
-        return {"detail": f"Candidate {candidateid} deleted"}
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Delete failed: {str(e)}")
-
-# --------------------------------------------------------Marketing--------------------------------------
-@app.get("/api/candidate/marketing", summary="Get all candidate marketing records")
-def read_all_marketing(page: int = Query(1, ge=1), limit: int = Query(100, ge=1, le=1000)):
-    return get_all_marketing_records(page, limit)
-
-@app.get("/api/candidate/marketing/{record_id}", summary="Get marketing record by ID")
-def read_marketing_record(record_id: int = Path(...)):
-    return get_marketing_by_id(record_id)
-
-@app.post("/api/candidate/marketing", response_model=CandidateMarketing, summary="Create candidate marketing record")
-def create_marketing_record(record: CandidateMarketingCreate):
-    return create_marketing(record)
-
-@app.put("/api/candidate/marketing/{record_id}", response_model=CandidateMarketing, summary="Update candidate marketing record")
-def update_marketing_record(record_id: int, record: CandidateMarketingCreate):
-    return update_marketing(record_id, record)
-
-@app.delete("/api/candidate/marketing/{record_id}", summary="Delete candidate marketing record")
-def delete_marketing_record(record_id: int):
-    return delete_marketing(record_id)
-
-# ------------------------------------------------------Placements--------------------------------------
-
-
-@app.get("/api/candidate/placements", summary="Get all placements (paginated)")
-def read_all_placements(page: int = Query(1, ge=1), limit: int = Query(100, ge=1, le=1000)):
-    return get_all_placements(page, limit)
-
-@app.get("/api/candidate/placements/{placement_id}", summary="Get placement by ID")
-def read_placement(placement_id: int = Path(...)):
-    return get_placement_by_id(placement_id)
-
-@app.post("/api/candidate/placements", response_model=CandidatePlacement, summary="Create a new placement")
-def create_new_placement(placement: CandidatePlacementCreate):
-    return create_placement(placement)
-
-@app.put("/api/candidate/placements/{placement_id}", response_model=CandidatePlacement, summary="Update placement")
-def update_existing_placement(placement_id: int, placement: CandidatePlacementCreate):
-    return update_placement(placement_id, placement)
-
-@app.delete("/api/candidate/placements/{placement_id}", summary="Delete placement")
-def delete_existing_placement(placement_id: int):
-    return delete_placement(placement_id)
-
-# ------------------------------------------------------- Leads -------------------------------------------------
-
-
-@app.get("/api/leads_new", summary="Get all leads (paginated)")
-async def get_all_leads(
-    page: int = Query(1, ge=1),
-    limit: int = Query(100, ge=1, le=1000)
-) -> Dict[str, Any]:
-    return fetch_all_leads_paginated(page, limit)
-
-
-@app.get("/api/leads_new/{lead_id}")
-def read_lead(lead_id: int = Path(...)) -> Dict[str, Any]:
-    return get_lead_by_id(lead_id)
-
-
-@app.post("/api/leads_new", response_model=Lead)
-def create_new_lead(lead: LeadCreate):
-    return create_lead(lead)
-
-
-@app.put("/api/leads_new/{lead_id}")
-def update_existing_lead(lead_id: int, lead: LeadCreate) -> Dict[str, Any]:
-    return update_lead(lead_id, lead)
-
-
-@app.delete("/api/leads_new/{lead_id}")
-def delete_existing_lead(lead_id: int) -> Dict[str, str]:
-    return delete_lead(lead_id)
-
-# -----------------------------------------------------------------------------------------------------
-
-# Temporary route to insert test data
-@app.get("/debug/insert-test-data")
-async def insert_test_data():
-    conn = get_connection()
-    cursor = conn.cursor()
-    cursor.execute("""
-    INSERT INTO newleads 
-    (full_name, email, phone, status) 
-    VALUES 
-    ('Test User', 'test@example.com', '1234567890', 'active')
-    """)
-    conn.commit()
-    cursor.close()
-    conn.close()
-    return {"message": "Test data inserted"}
-
-
-
+# # -----------------------------------------------------------------------------------------------------
 
 
 @app.get("/api/placements", response_model=List[RecentPlacement])
@@ -379,14 +165,10 @@ async def remove_interview(interview_id: int):
 
 
 
-
-
-
-
 # -------------------------------------------------------- IP -----------------------------------------
 
 
-@app.post("/vendor/request-demo")
+@app.post("/api/request-demo")
 async def create_vendor_request_demo(vendor: VendorCreate):
     try:
         vendor_data = vendor.dict()
@@ -434,6 +216,7 @@ async def register_google_user(request:Request,user: GoogleUserCreate):
 
     await insert_google_user_db(email=user.email, name=user.name, google_id=user.google_id)
     return {"message": "Google user registered successfully!"}
+    
 
 @app.post("/api/google_login/")
 @limiter.limit("15/minute")
@@ -451,7 +234,7 @@ async def login_google_user(request:Request,user: GoogleUserCreate):
         "google_id": existing_user['googleId'],
     }   
     
-    access_token = create_google_access_token(data=token_data)
+    access_token = await create_google_access_token(data=token_data)
     
     return {
         "access_token": access_token,
@@ -488,68 +271,6 @@ async def authenticate_user(uname: str, passwd: str):
     candidateid = candidate_info["candidateid"] if candidate_info else "Candidate ID not present"
     return {**user, "candidateid": candidateid}
 
-# Send email function
-def send_email_to_user(user_email: str, user_name: str):
-    from_email = os.getenv('EMAIL_USER')  # The "from" email (distributor)
-    to_admin_email = os.getenv('TO_RECRUITING_EMAIL')  # Admin email from environment variable
-    password = os.getenv('EMAIL_PASS')
-    smtp_server = os.getenv('SMTP_SERVER')
-    smtp_port = os.getenv('SMTP_PORT')
-
-    # Email content for the user
-    user_html_content = f"""
-    <html>
-        <body>
-            <p>Dear {user_name},</p>
-            <p>Thank you for registering with us. We are pleased to inform you that our recruiting team will reach out to you shortly.</p>
-            <p>Best regards,<br>Recruitment Team</p>
-        </body>
-    </html>
-    """
-
-    # Email content for the admin
-    admin_html_content = f"""
-    <html>
-        <body>
-            <p>Hello Admin,</p>
-            <p>A new user has registered on the website. Please review their details and provide access.</p>
-            <p><strong>User Details:</strong></p>
-            <ul>
-                <li>Name: {user_name}</li>
-                <li>Email: {user_email}</li>
-            </ul>
-            <p>Best regards,<br>System Notification</p>
-        </body>
-    </html>
-    """
-
-    try:
-        # Establish the connection with the email server
-        server = smtplib.SMTP(smtp_server, int(smtp_port))
-        server.starttls()
-        server.login(from_email, password)
-
-        # Send email to the user
-        user_msg = MIMEMultipart()
-        user_msg['From'] = from_email
-        user_msg['To'] = user_email
-        user_msg['Subject'] = 'Registration Successful - Recruiting Team will Reach Out'
-        user_msg.attach(MIMEText(user_html_content, 'html'))
-        server.sendmail(from_email, user_email, user_msg.as_string())
-
-        # Send email to the admin
-        admin_msg = MIMEMultipart()
-        admin_msg['From'] = from_email
-        admin_msg['To'] = to_admin_email
-        admin_msg['Subject'] = 'New User Registration Notification'
-        admin_msg.attach(MIMEText(admin_html_content, 'html'))
-        server.sendmail(from_email, to_admin_email, admin_msg.as_string())
-
-        # Close the server
-        server.quit()
-
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f'Error while sending emails: {e}')
 
 def clean_input_fields(user_data: UserRegistration):
     """Convert empty strings and format datetimes for MySQL"""
@@ -570,6 +291,7 @@ def clean_input_fields(user_data: UserRegistration):
 @app.post("/api/signup")
 @limiter.limit("15/minute")
 async def register_user(request:Request,user: UserRegistration):
+    user.uname = user.uname.lower().strip()
     # Check if user exists
     existing_user = await get_user_by_username(user.uname)
     if existing_user:
@@ -582,11 +304,12 @@ async def register_user(request:Request,user: UserRegistration):
     hashed_password = md5_hash(user.passwd)
     # fullname = f"{user.firstname or ''} {user.lastname or ''}".strip(),
     fullname = user.fullname or f"{user.firstname or ''} {user.lastname or ''}".strip()
+    fullname = fullname.lower()
+    user.fullname = fullname
+    leads_full_name = f"{user.firstname or ''} {user.lastname or ''}".strip()
     # print(" Full name constructed:", fullname)  # <---- Add this line
 
   
-    
-
     await insert_user(
     uname=user.uname,
     passwd=hashed_password,
@@ -596,8 +319,7 @@ async def register_user(request:Request,user: UserRegistration):
     instructor=user.instructor,
     override=user.override,
     lastlogin=user.lastlogin,
-    logincount=user.logincount,
-    # fullname=user.fullname,
+    logincount=user.logincount,   
     fullname=fullname,
     phone=user.phone,
     address=user.address,
@@ -606,11 +328,11 @@ async def register_user(request:Request,user: UserRegistration):
     country=user.country,
     message=user.message,
     registereddate=user.registereddate,
-    level3date=user.level3date,
-    visastatus=user.visastatus,
+    level3date=user.level3date,    
+    visa_status=user.visa_status,
     experience=user.experience,
     education=user.education,
-    referred_by=user.referred_by,
+    referby=user.referby,
     candidate_info={  # optional dict
         'name': user.fullname,
         'enrolleddate': user.registereddate,
@@ -623,13 +345,22 @@ async def register_user(request:Request,user: UserRegistration):
         }
     )
 
+    await insert_lead_new(
+    full_name=leads_full_name,
+    phone=user.phone,
+    email=user.uname,
+    address=user.address,
+    workstatus=None,  # If available, pass user.workstatus
+    status="Open",
+    secondary_email=None,  # Or user.secondaryemail if you collect it
+    secondary_phone=None,  # Or user.secondaryphone
+    closed_date=None,
+    notes=None
+    )
 
     # Send confirmation email to the user and notify the admin
-    send_email_to_user(user_email=user.uname, user_name=user.fullname)
-
+    send_email_to_user(user_email=user.uname, user_name=user.fullname, user_phone=user.phone)
     return {"message": "User registered successfully. Confirmation email sent to the user and notification sent to the admin."}
-
-
 
 
 @app.post("/api/login", response_model=Token)
@@ -800,43 +531,26 @@ async def get_recordings(request:Request,course: str = None, batchid: int = None
         raise HTTPException(status_code=500, detail=str(e))
 
 @app.post("/api/contact")
-
 async def contact(user: ContactForm):
-    
-    
-
-
-    await user_contact(
-        # name=f"{user.firstName} {user.lastName}",
-        name=f"{user.firstName} {user.lastName}",
+        # Send emails
+    send_contact_emails(
+        first_name=user.firstName,
+        last_name=user.lastName,
         email=user.email,
         phone=user.phone,
-        message=user.message        
-        )
-    def sendEmail():
-        from_Email = os.getenv('EMAIL_USER')
-        password = os.getenv('EMAIL_PASS')
-        to_email = os.getenv('TO_RECRUITING_EMAIL')
-        smtp_server = os.getenv('SMTP_SERVER')
-        smtp_port = os.getenv('SMTP_PORT')
-        html_content = ContactMail_HTML_templete(f"{user.firstName} {user.lastName}",user.email,user.phone,user.message)
-        msg = MIMEMultipart()
-        msg['From'] = from_Email
-        msg['To'] = to_email
-        msg['Subject'] = 'WBL Contact lead generated'
-        msg.attach(MIMEText(html_content, 'html'))
-        try:
-            # server = smtplib.SMTP('smtp.gmail.com',587)
-            server = smtplib.SMTP(smtp_server, int(smtp_port))
-            server.starttls()
-            server.login(from_Email,password)
-            text = msg.as_string()
-            server.sendmail(from_Email,to_email,text)
-            server.quit()
-        except Exception as e:
-            raise HTTPException(status_code=500, detail='Erro while sending the mail to recruiting teams')
-    sendEmail()
-    return {"detail": "Message Sent Successfully"}
+        message=user.message
+    )
+    
+    # Save to database
+    full_name = f"{user.firstName} {user.lastName}"
+    await user_contact(
+        full_name=full_name,
+        email=user.email,
+        phone=user.phone,
+        message=user.message
+    )
+    return {"detail": "Message sent successfully"}
+
 
 @app.get("/api/coursecontent")
 def get_course_content():
@@ -871,9 +585,8 @@ async def reset_password(data: ResetPassword):
 
 
 # ...................................NEW INNOVAPATH......................................
-
-@app.get("/api/candidate_marketing/", response_model=List[CandidateMarketing])
-async def get_candidate_marketing(
+@app.get("/api/talent_search", response_model=List[TalentSearch])
+async def get_talent_search(
     role: Optional[str] = None,
     experience: Optional[int] = None,
     location: Optional[str] = None,
@@ -897,6 +610,3 @@ async def get_candidate_marketing(
         raise e
     except Exception as e:
         raise HTTPException(status_code=500, detail=f"Internal server error: {str(e)}")
-    
-  
-
